move continuity screenshot asserter to dedicated helper module
