+++ e2e-pw/src/oss/poms/modal/index.ts
@@ -8,10 +8,7 @@ import { ModalGroupActionsPom } from "./group-actions";
 import { ModalImaAsVideoControlsPom } from "./imavid-controls";
 import { Looker3DControlsPom } from "./looker-3d-controls";
 import { ModalSidebarPom } from "./modal-sidebar";
-import {
-  compareLocatorScreenshotToBuffer,
-  SampleCanvasPom,
-} from "./sample-canvas";
+import { SampleCanvasPom } from "./sample-canvas";
 import { ModalVideoControlsPom } from "./video-controls";
 
 const SAMPLE_LOAD_TIMEOUT = Duration.Seconds(20);
@@ -313,12 +310,6 @@ export class ModalPom {
       { timeout: SAMPLE_LOAD_TIMEOUT }
     );
   }
-
-  async captureStableMain2DRendererScreenshot() {
-    await this.groupLooker.waitFor({ state: "visible" });
-    await this.sampleCanvas.moveMouseToViewportEdge();
-    return this.groupLooker.screenshot();
-  }
 }
 
 class ModalAsserter {
@@ -350,19 +341,6 @@ class ModalAsserter {
   async verify3dRendererVisible() {
     await expect(this.modalPom.looker3d).toBeVisible();
   }
-
-  async verifyMain2RendererMatchesScreenshotBuffer(expectedScreenshot: Buffer) {
-    await expect(this.modalPom.groupLooker).toBeVisible();
-    await this.modalPom.sampleCanvas.tooltip.assert.isVisible(false);
-    await this.modalPom.sampleCanvas.toolbar.assert.isVisible(false);
-    await this.modalPom.sampleCanvas.moveMouseToViewportEdge();
-
-    await compareLocatorScreenshotToBuffer(
-      this.modalPom.groupLooker,
-      expectedScreenshot
-    );
-  }
-
   async verifySelectionCount(n: number) {
     const action = this.modalPom.locator.getByTestId("action-manage-selected");
