+++ e2e-pw/src/oss/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./commands";
 export * from "./dom";
 export * from "./duration";
+export * from "./screenshot";
 export * from "./str";
